tabele delajo, potreben še update če se spremeni vrednost izdelka, pogledat qr kodo, narediti izbris seznama z on long klick, popravit menije,...

+++ src/edu/pernat/shopinglist/android/razredi/NovSeznamArtiklov.java
@@ -58,7 +58,7 @@ public class NovSeznamArtiklov {
 	private double skupnaCena;
 	private String imeSeznama;
 	private int stOznacenih;
-	private int idBaze;
+	private long idBaze;
 	
 	public int getStOznacenih() {
 		prestejOznacene();
@@ -129,12 +129,12 @@ public class NovSeznamArtiklov {
 		stOznacenih=vmesni;
 	}
 	
-	public int getIdBaze() {
+	public long getIdBaze() {
 		return idBaze;
 	}
 
 
-	public void setIdBaze(int idBaze) {
+	public void setIdBaze(long idBaze) {
 		this.idBaze = idBaze;
 	}
